fix: mongo being used client side
